REM (Retrieving Effectively from Memory) model

Shiffrin & Steyvers' (1997) REM adapted to cross-situational word learning. Word/object identity is a length-w feature vector drawn from a geometric base-rate distribution (g); each trial stores one episodic trace per candidate word-object pairing present, built by imperfectly copying the pair's true features (each feature independently stored with probability u, and if stored, correct with probability c, else a fresh random draw). At test, a candidate pairing's strength is the summed likelihood ratio of every trace given the true (fully-known) probe features, REM's odds-based alternative to minerva2()'s dot-product activation.

Usage

rem(g, u, c, w = 12)

Arguments

g

Base-rate feature-value distribution parameter (larger = more variable/distinctive feature values, in (0, 1))

u

Per-feature storage probability at study, in (0, 1)

c

Probability a stored feature is copied correctly (vs. replaced by a random, possibly-confusable value), in (0, 1)

w

Number of features per item (fixed hyperparameter, not a free cognitive parameter)

Value

An object of class xslMod

Details

Performance note: unlike minerva2() (one trace per trial) this model stores one trace per candidate (word, object) pair per trial – on a highly ambiguous corpus (many words/objects per utterance) memory size grows combinatorially rather than linearly with trial count, and both encoding and the voc_sz * ref_sz test pass scale with that trace count, making rem() the most expensive of this package's memory models on cluttered/naturalistic data. w is a fixed representational hyperparameter (not usually treated as a free cognitive parameter), not included by default in this package's DEoptim registry bounds – see xsl_model_registry().

Examples

mod <- rem(g = 0.4, u = 0.3, c = 0.7, w = 12)
xsl_run(mod, get_example_ambiguous_condition())
#> $fits
#> $fits[[1]]
#> $fits[[1]]$sims
#> NULL
#> 
#> $fits[[1]]$responses
#>                [,1]         [,2]         [,3]         [,4]
#>   [1,] 9.316867e-01 3.681924e-02 7.375293e-02 8.398092e-01
#>   [2,] 1.078624e-01 1.930415e-02 3.413415e-03 5.147715e-01
#>   [3,] 9.131010e-01 2.598595e-02 2.276025e-03 9.946749e-02
#>   [4,] 5.267167e-03 2.935090e-01 6.868332e-02 2.580907e-01
#>   [5,] 1.766570e-01 4.846060e-01 4.110811e-01 2.802045e-01
#>   [6,] 1.745762e-01 8.895385e-01 1.621693e-01 1.358150e-01
#>   [7,] 8.384505e-01 3.110007e-02 7.295689e-01 8.755929e-01
#>   [8,] 4.191625e-01 2.492499e-02 6.820770e-01 7.912698e-02
#>   [9,] 2.291809e-01 8.072550e-01 2.118386e-01 1.234581e-01
#>  [10,] 3.024591e-02 1.125971e-01 8.837331e-01 5.894406e-01
#>  [11,] 4.600988e-02 6.636919e-01 9.725003e-01 2.915442e-04
#>  [12,] 8.373622e-01 1.188756e-01 9.027349e-01 1.194930e-01
#>  [13,] 7.695316e-04 9.998448e-01 3.878665e-02 1.761717e-02
#>  [14,] 7.717479e-01 2.678946e-03 9.975512e-01 3.155126e-04
#>  [15,] 1.472549e-01 3.166501e-01 1.314942e-01 2.974852e-01
#>  [16,] 3.852380e-01 5.740466e-01 3.004857e-01 3.184697e-01
#>  [17,] 6.257519e-02 7.418078e-01 6.691782e-03 7.293883e-01
#>  [18,] 3.158923e-01 6.404622e-01 5.455255e-02 1.673127e-01
#>  [19,] 1.136704e-01 1.677852e-01 2.454180e-01 3.848250e-01
#>  [20,] 5.539981e-02 1.721205e-01 2.770891e-01 6.517739e-01
#>  [21,] 6.285131e-02 4.473949e-02 1.109517e-01 2.115075e-01
#>  [22,] 4.711296e-01 7.574852e-01 1.589706e-01 2.983790e-02
#>  [23,] 1.953173e-01 3.912938e-01 4.883422e-01 3.857498e-01
#>  [24,] 8.075376e-02 1.314815e-01 3.756761e-01 1.438618e-02
#>  [25,] 2.767195e-02 1.882274e-03 9.831734e-01 3.708870e-01
#>  [26,] 8.917720e-03 2.984363e-02 9.924313e-01 7.621341e-03
#>  [27,] 4.971995e-01 8.930973e-02 7.349490e-01 6.569596e-01
#>  [28,] 1.330832e-01 6.913114e-01 3.762760e-01 2.007381e-01
#>  [29,] 2.227354e-01 1.166289e-01 3.426371e-01 8.105700e-01
#>  [30,] 8.637978e-01 1.603212e-01 3.119780e-01 3.864494e-02
#>  [31,] 1.525629e-02 2.663507e-01 9.880356e-01 9.244780e-01
#>  [32,] 8.556951e-01 6.280949e-01 2.533953e-01 1.942462e-01
#>  [33,] 6.255472e-03 9.235207e-01 4.133768e-01 1.241501e-01
#>  [34,] 5.239867e-01 1.863197e-01 8.392453e-02 4.751479e-01
#>  [35,] 2.468404e-01 8.253648e-01 6.327102e-01 8.338353e-02
#>  [36,] 4.814274e-01 3.105766e-01 6.006296e-02 3.200760e-01
#>  [37,] 1.873566e-01 2.926738e-02 3.428742e-02 7.554523e-03
#>  [38,] 3.101840e-02 1.920227e-01 7.618852e-01 1.239586e-02
#>  [39,] 1.681760e-03 1.238533e-04 2.982623e-03 3.292755e-01
#>  [40,] 6.381041e-02 2.530368e-03 9.989372e-01 7.356566e-04
#>  [41,] 9.146924e-01 3.483989e-02 1.830371e-01 1.982734e-02
#>  [42,] 2.575084e-01 1.702201e-01 2.930678e-02 3.647596e-02
#>  [43,] 2.276036e-01 1.610441e-01 4.802976e-02 9.709171e-01
#>  [44,] 3.546196e-01 1.463016e-01 7.450373e-02 3.261088e-02
#>  [45,] 9.176523e-01 7.274228e-02 1.438880e-01 4.124556e-01
#>  [46,] 1.833633e-01 3.001908e-01 2.166370e-01 3.422148e-01
#>  [47,] 8.627125e-01 4.150035e-01 4.958440e-01 2.278999e-01
#>  [48,] 1.457818e-01 1.476205e-01 8.786089e-02 5.776650e-01
#>  [49,] 5.677587e-01 1.376204e-01 8.343299e-01 1.626773e-01
#>  [50,] 2.441434e-01 4.046564e-01 1.940666e-01 1.144414e-01
#>  [51,] 4.877968e-02 4.053720e-02 4.060341e-02 1.276614e-01
#>  [52,] 7.242092e-02 3.470028e-01 8.165013e-02 9.684925e-01
#>  [53,] 6.909278e-02 2.164393e-01 8.280046e-01 9.682521e-02
#>  [54,] 7.182135e-01 3.021026e-02 1.687997e-02 6.882800e-01
#>  [55,] 3.417306e-01 2.730814e-02 6.229828e-03 2.061676e-02
#>  [56,] 5.707792e-01 8.132389e-03 5.956771e-02 5.523359e-01
#>  [57,] 5.882352e-01 4.003733e-01 1.554856e-01 8.465021e-01
#>  [58,] 8.163322e-03 8.068547e-01 2.877695e-02 4.385056e-02
#>  [59,] 2.940238e-03 4.846555e-03 9.714222e-01 2.603945e-02
#>  [60,] 2.129010e-03 8.867217e-01 1.332930e-02 9.338567e-04
#>  [61,] 5.643737e-02 1.780039e-01 1.657443e-02 5.218892e-01
#>  [62,] 3.079222e-01 2.396616e-01 6.837350e-03 4.010683e-01
#>  [63,] 2.617967e-01 2.642133e-01 2.482863e-01 5.324268e-01
#>  [64,] 1.298208e-01 4.151776e-02 2.240579e-02 1.134704e-01
#>  [65,] 2.906318e-01 1.650819e-01 4.304088e-01 4.084244e-01
#>  [66,] 6.350099e-01 2.870364e-04 5.455808e-01 1.614722e-03
#>  [67,] 1.577072e-01 4.457839e-01 7.652906e-01 5.554122e-03
#>  [68,] 6.744330e-01 1.454673e-01 3.026769e-01 9.697821e-01
#>  [69,] 4.464409e-01 3.989972e-01 5.669754e-01 2.422904e-01
#>  [70,] 8.002144e-02 5.499887e-03 5.865503e-01 1.484169e-01
#>  [71,] 1.076340e-01 1.569420e-01 3.402060e-01 2.818204e-01
#>  [72,] 8.877733e-01 1.017391e-01 1.982573e-02 7.765330e-01
#>  [73,] 7.357930e-02 2.809628e-01 5.264714e-01 2.070628e-01
#>  [74,] 8.328736e-01 9.299888e-03 2.386356e-02 5.104529e-03
#>  [75,] 1.168797e-02 9.606962e-01 7.938428e-02 7.476951e-02
#>  [76,] 2.940669e-01 1.667627e-02 9.162172e-02 9.897981e-01
#>  [77,] 3.130845e-01 2.148188e-01 3.318201e-01 5.602832e-01
#>  [78,] 1.557949e-04 9.971962e-01 6.712731e-01 6.172471e-01
#>  [79,] 1.386881e-02 4.015442e-03 4.612253e-03 8.604743e-01
#>  [80,] 3.252072e-02 6.007906e-03 3.353627e-02 8.336102e-02
#>  [81,] 2.535430e-01 2.413956e-01 2.400013e-01 1.385002e-01
#>  [82,] 4.656635e-02 1.861941e-01 5.788076e-01 8.376905e-01
#>  [83,] 1.298357e-01 8.541758e-02 3.062124e-01 5.549954e-02
#>  [84,] 1.470621e-02 1.070013e-01 3.730155e-01 5.375887e-01
#>  [85,] 2.772819e-01 1.539979e-03 3.478953e-02 1.640391e-01
#>  [86,] 1.804708e-01 2.948229e-02 5.851815e-02 1.714755e-01
#>  [87,] 3.451459e-01 6.503334e-01 8.856787e-01 4.533698e-01
#>  [88,] 4.087863e-01 2.705520e-01 3.247459e-01 6.130578e-01
#>  [89,] 4.021833e-03 2.819714e-02 9.999112e-01 3.764273e-02
#>  [90,] 6.421055e-01 5.244687e-01 6.319673e-01 9.894184e-03
#>  [91,] 8.824151e-03 2.737977e-02 8.307495e-03 2.261802e-01
#>  [92,] 1.507943e-02 1.671690e-02 3.081719e-02 9.715398e-01
#>  [93,] 6.321316e-01 4.800617e-01 2.222773e-01 1.868998e-01
#>  [94,] 1.901873e-01 1.557978e-01 5.745252e-01 1.816531e-01
#>  [95,] 9.530591e-02 2.353845e-01 5.523226e-01 9.679895e-02
#>  [96,] 5.235226e-02 7.946311e-01 5.685013e-01 3.714260e-01
#>  [97,] 5.757771e-01 1.338510e-01 1.069423e-03 4.687918e-01
#>  [98,] 2.396965e-01 3.123795e-02 6.631262e-01 3.728520e-01
#>  [99,] 3.450841e-01 1.267355e-01 5.693156e-01 3.572850e-02
#> [100,] 1.760822e-01 3.245966e-01 5.245168e-01 1.349813e-02
#> [101,] 4.824615e-02 5.812394e-01 8.658854e-01 1.208348e-01
#> [102,] 4.475546e-01 8.564416e-01 8.812126e-01 5.109326e-02
#> [103,] 2.596984e-01 5.722843e-02 2.570402e-01 5.260195e-02
#> [104,] 4.981817e-04 2.627172e-04 3.983330e-02 9.977470e-01
#> [105,] 1.003522e-02 1.066424e-04 4.370436e-03 9.974412e-01
#> [106,] 9.589410e-02 8.049247e-02 1.734618e-03 9.992372e-01
#> [107,] 7.427768e-01 2.196443e-02 7.490191e-01 2.447756e-02
#> [108,] 2.846772e-01 3.517145e-02 3.873876e-01 3.434883e-01
#> [109,] 9.060662e-01 1.410001e-01 7.230321e-02 7.476850e-01
#> [110,] 8.624117e-01 8.844458e-01 5.312232e-01 2.625225e-01
#> [111,] 3.970336e-01 1.754590e-01 2.914957e-01 3.056017e-01
#> [112,] 4.047540e-02 8.872623e-01 3.104653e-02 1.137573e-01
#> [113,] 2.824670e-01 5.027803e-03 3.464783e-01 1.121314e-04
#> [114,] 1.022615e-04 8.755775e-01 3.910462e-02 9.930690e-01
#> [115,] 7.355564e-01 3.071828e-01 1.448367e-01 4.178417e-01
#> [116,] 3.478190e-02 7.007603e-02 3.736670e-02 9.864559e-01
#> [117,] 1.128801e-02 1.502066e-01 9.765878e-01 3.043958e-01
#> [118,] 8.908799e-01 4.306695e-02 4.498761e-01 4.315682e-02
#> [119,] 6.322107e-01 1.274972e-01 1.896253e-02 3.685760e-01
#> [120,] 6.583818e-03 9.228150e-01 4.032160e-01 5.398552e-02
#> [121,] 9.733123e-01 4.357994e-02 1.149778e-01 3.712550e-02
#> [122,] 9.104841e-01 1.363033e-01 2.021383e-01 4.699369e-01
#> [123,] 2.042140e-02 9.565148e-01 3.035485e-02 5.308559e-02
#> [124,] 4.949697e-01 6.701574e-01 9.708765e-01 2.560483e-01
#> [125,] 3.546903e-02 3.137691e-01 7.954621e-01 1.762963e-03
#> [126,] 1.246071e-02 1.209926e-01 8.742544e-01 1.720796e-01
#> [127,] 8.195230e-02 9.517782e-02 9.276982e-01 1.647578e-01
#> [128,] 1.179559e-01 9.871126e-01 3.709365e-01 1.984661e-02
#> [129,] 6.102674e-01 2.674485e-01 5.957937e-01 9.159405e-02
#> [130,] 3.521950e-01 1.317480e-01 4.199701e-01 3.416618e-01
#> [131,] 3.086931e-02 3.634215e-01 4.023715e-01 1.453697e-01
#> [132,] 1.011901e-01 3.624557e-01 3.320228e-01 4.164946e-01
#> [133,] 8.967059e-02 3.323288e-01 2.674867e-02 7.471827e-01
#> [134,] 3.533411e-01 5.719321e-03 5.604667e-01 1.005933e-02
#> [135,] 9.716038e-02 7.606164e-01 1.633786e-01 5.351984e-02
#> [136,] 6.901939e-01 1.613698e-01 3.902235e-01 4.703155e-01
#> [137,] 1.172700e-01 4.939732e-02 9.415377e-01 1.283251e-02
#> [138,] 1.734273e-01 3.114795e-02 9.822811e-01 1.979195e-01
#> [139,] 4.577526e-01 3.083970e-04 9.921473e-01 1.067621e-02
#> [140,] 2.196565e-01 7.582276e-01 3.891690e-01 1.037941e-01
#> [141,] 1.676923e-01 4.985764e-01 4.426807e-01 2.005417e-02
#> [142,] 4.092248e-02 6.319292e-01 2.784300e-03 8.145643e-01
#> [143,] 6.528365e-01 3.399361e-01 8.031290e-01 5.088131e-04
#> [144,] 8.661535e-02 1.558304e-01 5.628889e-01 3.534764e-01
#> [145,] 1.983126e-02 1.788602e-01 3.149325e-02 1.156641e-01
#> [146,] 2.026124e-01 2.527500e-01 6.745076e-02 4.969223e-01
#> [147,] 7.148960e-01 6.643572e-02 1.522342e-01 1.238874e-01
#> [148,] 3.146675e-01 3.647130e-01 3.357293e-01 4.205172e-01
#> [149,] 6.740149e-01 2.046369e-01 2.456470e-01 1.769221e-01
#> [150,] 9.993932e-01 3.533280e-04 9.688559e-04 3.137394e-02
#> [151,] 3.351711e-01 2.065248e-01 3.021201e-01 3.617228e-01
#> [152,] 1.267974e-01 9.922566e-01 1.969641e-01 1.000178e-01
#> [153,] 2.490996e-01 1.828238e-01 4.622114e-03 3.387342e-02
#> [154,] 3.183863e-01 2.371130e-01 1.871391e-01 2.770613e-01
#> [155,] 1.366188e-01 8.900339e-01 6.560410e-02 8.870125e-02
#> [156,] 1.415670e-01 2.695254e-02 8.567219e-01 9.188399e-04
#> [157,] 4.628672e-03 8.732745e-01 9.608852e-01 1.116216e-02
#> [158,] 7.962500e-02 5.899190e-01 9.677314e-01 9.781802e-03
#> [159,] 9.992117e-01 6.323295e-03 9.007938e-03 9.080849e-01
#> [160,] 7.155956e-01 1.501830e-01 2.226849e-01 4.761839e-01
#> [161,] 1.432144e-01 1.955960e-03 9.535082e-01 1.242214e-02
#> [162,] 2.147291e-02 8.335290e-03 6.042293e-01 3.479166e-03
#> [163,] 2.336724e-01 2.755620e-02 3.091413e-01 3.829711e-01
#> [164,] 7.108365e-01 1.908821e-01 9.770966e-01 1.910639e-01
#> [165,] 2.649435e-01 4.492064e-01 7.040817e-02 5.590934e-02
#> [166,] 4.263362e-01 7.882365e-03 9.871356e-01 8.394363e-03
#> [167,] 6.275268e-01 9.951816e-01 3.602025e-03 5.950832e-01
#> [168,] 6.722943e-01 6.317134e-01 9.775942e-01 9.767655e-02
#> [169,] 5.357067e-01 3.878189e-01 4.125762e-02 7.348687e-01
#> [170,] 1.044978e-02 9.997275e-01 3.669681e-03 9.815153e-01
#> [171,] 8.198227e-01 2.179248e-01 4.856467e-01 1.116756e-01
#> [172,] 1.133251e-01 3.841058e-01 2.573675e-02 3.990904e-01
#> [173,] 5.035068e-01 5.689015e-01 1.197456e-01 9.257757e-01
#> [174,] 2.267580e-01 4.462481e-02 7.403318e-02 9.495165e-01
#> [175,] 5.463755e-04 9.972782e-01 1.774853e-01 9.838170e-04
#> [176,] 3.814771e-01 1.798694e-02 1.495743e-01 8.259264e-01
#> [177,] 2.139656e-01 8.726501e-02 2.502514e-01 4.925651e-01
#> [178,] 2.856256e-01 4.921650e-02 6.013101e-01 2.097148e-02
#> [179,] 1.790041e-02 9.915703e-01 5.567427e-02 1.517327e-02
#> [180,] 9.807483e-01 4.495364e-03 8.481601e-03 1.481922e-03
#> [181,] 3.227050e-02 1.851029e-01 8.223346e-01 9.919385e-01
#> [182,] 2.079532e-01 1.545531e-01 8.356458e-01 7.240144e-02
#> [183,] 9.382397e-01 1.144880e-01 2.028480e-02 2.230393e-02
#> [184,] 5.767573e-01 1.629581e-01 5.515569e-02 5.819933e-02
#> [185,] 2.076920e-01 2.292865e-01 3.043124e-01 1.465139e-01
#> [186,] 4.623698e-01 4.671238e-01 3.165057e-01 8.854025e-02
#> [187,] 4.364295e-01 3.574378e-02 3.358159e-01 6.370455e-01
#> [188,] 9.154450e-01 3.133680e-02 3.627205e-01 9.504246e-03
#> [189,] 5.525262e-02 6.729633e-01 1.929647e-02 4.197209e-01
#> [190,] 4.195932e-01 2.505078e-01 9.897266e-02 3.066529e-01
#> [191,] 2.681027e-01 5.148565e-01 2.487660e-01 7.846073e-01
#> [192,] 8.922135e-01 4.451373e-02 6.381706e-01 1.803903e-02
#> [193,] 2.595672e-02 2.249816e-01 3.114642e-01 3.944234e-02
#> [194,] 9.776748e-01 4.140887e-01 1.074571e-01 2.129391e-01
#> [195,] 2.270169e-01 5.517914e-02 5.037542e-02 2.927647e-01
#> [196,] 4.024078e-03 1.100137e-03 8.604096e-01 9.722934e-04
#> [197,] 7.021478e-02 3.444359e-01 7.871570e-02 6.167253e-02
#> [198,] 1.687116e-01 8.391453e-01 6.554266e-03 1.769335e-01
#> [199,] 4.029418e-01 1.184476e-01 5.885094e-01 8.231939e-03
#> [200,] 1.221440e-01 9.917571e-01 9.128819e-01 4.704162e-03
#> [201,] 6.988860e-01 1.526926e-02 1.869435e-01 3.741456e-02
#> [202,] 2.659620e-01 2.391357e-02 2.760195e-01 1.813751e-01
#> [203,] 2.918592e-02 8.613482e-01 9.491260e-01 4.454078e-02
#> [204,] 6.173406e-02 8.016452e-01 5.870628e-03 1.626020e-01
#> [205,] 9.914144e-01 1.090990e-01 5.204909e-01 7.428033e-02
#> [206,] 6.913318e-01 8.275834e-01 1.767149e-01 2.189115e-01
#> [207,] 5.970661e-02 1.400832e-01 8.660974e-01 6.902985e-01
#> [208,] 7.173456e-02 3.784231e-01 3.882593e-02 9.854284e-02
#> [209,] 6.330763e-01 6.642573e-01 9.857565e-01 3.753724e-02
#> [210,] 7.702820e-02 3.698735e-02 9.952611e-01 5.734466e-03
#> [211,] 2.304464e-01 2.104158e-01 1.331396e-01 2.652754e-01
#> [212,] 9.819351e-01 3.713560e-02 7.924623e-02 2.113940e-03
#> [213,] 3.449770e-02 9.495190e-01 9.602784e-01 1.256383e-01
#> [214,] 1.372219e-01 5.648189e-02 1.044895e-02 2.922165e-01
#> [215,] 7.923551e-01 8.158423e-01 1.962086e-01 8.687463e-03
#> [216,] 1.965713e-01 3.487767e-01 2.272919e-01 9.317382e-02
#> [217,] 9.103546e-01 5.542156e-02 7.010584e-01 5.906504e-02
#> [218,] 2.129861e-01 6.998006e-01 9.167552e-01 5.435339e-02
#> [219,] 6.439177e-01 2.916218e-01 9.139065e-02 9.451765e-01
#> [220,] 2.789905e-02 4.745598e-01 7.538881e-01 3.507864e-02
#> [221,] 7.731255e-01 2.006585e-02 5.470906e-01 7.672458e-01
#> [222,] 7.497076e-01 6.725500e-03 1.301363e-01 4.385570e-01
#> [223,] 1.156513e-02 3.640246e-01 4.966901e-01 5.572426e-02
#> [224,] 4.818342e-01 5.120825e-01 1.996536e-02 2.822794e-01
#> [225,] 4.792987e-01 4.622243e-01 1.861274e-02 4.010324e-01
#> [226,] 1.986064e-01 3.773883e-02 1.801555e-02 9.777546e-01
#> [227,] 8.615038e-01 8.354267e-04 7.516578e-01 4.397060e-03
#> [228,] 3.430686e-01 2.186445e-01 5.543066e-01 5.043554e-01
#> [229,] 5.020037e-01 5.324487e-01 6.616494e-01 1.184358e-02
#> [230,] 1.287289e-01 1.832937e-01 3.733212e-01 6.138149e-01
#> [231,] 3.492763e-02 8.195565e-01 8.297798e-02 9.250606e-03
#> [232,] 8.998778e-01 5.323165e-05 8.065797e-02 1.200561e-04
#> [233,] 3.680762e-01 4.154213e-02 8.719530e-02 4.973496e-01
#> [234,] 6.049203e-01 1.983760e-01 2.141540e-01 6.856394e-01
#> [235,] 2.277454e-01 1.693227e-01 1.777194e-02 7.478247e-01
#> [236,] 6.596476e-01 5.799491e-02 9.947374e-01 1.258710e-01
#> [237,] 2.960224e-01 2.875725e-01 4.484812e-02 3.456049e-01
#> [238,] 3.390835e-01 1.754184e-02 6.150735e-02 6.792535e-01
#> [239,] 1.570508e-01 7.769192e-04 9.954640e-01 2.944529e-02
#> [240,] 2.041438e-01 1.178288e-01 4.844484e-01 9.762540e-01
#> [241,] 1.500045e-01 1.511186e-01 1.049819e-01 2.153348e-01
#> [242,] 6.613174e-02 9.356143e-02 7.716534e-01 3.628935e-01
#> [243,] 2.045750e-05 3.786333e-02 5.549423e-01 7.958035e-03
#> [244,] 1.806210e-01 5.074223e-01 3.750564e-01 3.333727e-01
#> [245,] 2.654087e-02 9.557582e-01 3.692746e-03 3.421581e-02
#> [246,] 1.362781e-01 4.761677e-03 7.071689e-01 5.146986e-04
#> [247,] 3.414641e-01 1.089856e-01 6.134038e-01 3.818406e-01
#> [248,] 9.380856e-01 1.653322e-01 4.748596e-01 2.816530e-02
#> [249,] 2.241704e-02 3.145237e-03 2.570841e-01 2.126500e-01
#> [250,] 1.708550e-01 4.380652e-01 8.665116e-01 3.259678e-03
#> [251,] 8.914944e-01 4.390340e-01 1.483535e-02 1.287878e-01
#> [252,] 1.079669e-02 3.016563e-01 1.892208e-01 2.404516e-01
#> [253,] 7.186886e-01 1.428088e-01 1.014307e-01 3.062274e-01
#> [254,] 9.872412e-01 2.022183e-01 1.340283e-03 2.870236e-01
#> [255,] 3.556969e-02 2.582890e-01 7.298850e-02 1.098312e-01
#> [256,] 7.524287e-01 1.608378e-02 3.486423e-02 9.234350e-01
#> [257,] 6.414941e-01 2.474727e-01 2.491372e-02 9.882943e-01
#> [258,] 6.453259e-02 9.025346e-02 9.628243e-01 2.387369e-01
#> [259,] 1.518235e-06 6.027953e-06 9.999869e-01 1.040975e-04
#> [260,] 5.809428e-01 1.615294e-01 2.835013e-01 1.510206e-02
#> [261,] 8.902790e-02 7.670627e-01 9.974683e-02 1.517197e-01
#> [262,] 2.938585e-01 6.404404e-02 4.031224e-01 1.861784e-02
#> [263,] 3.315779e-02 6.948182e-02 9.999773e-01 9.637756e-01
#> [264,] 1.176057e-01 2.935283e-01 9.805141e-01 1.217854e-01
#> [265,] 4.698950e-01 1.956548e-01 2.596126e-01 8.966029e-02
#> [266,] 4.324461e-01 1.808116e-02 9.689142e-01 8.882563e-03
#> [267,] 1.638000e-01 2.664339e-01 3.640050e-01 5.089339e-01
#> [268,] 5.072274e-01 4.260430e-01 1.724169e-01 1.267791e-01
#> [269,] 3.394402e-03 4.015737e-02 2.186890e-02 9.956417e-01
#> [270,] 5.265322e-01 6.827725e-01 4.287120e-01 5.656352e-01
#> [271,] 7.594997e-02 8.591655e-02 6.803020e-02 5.383356e-01
#> [272,] 5.548255e-02 8.053402e-01 7.331805e-01 4.438822e-02
#> [273,] 2.022436e-01 2.072910e-01 1.407813e-01 2.186182e-01
#> [274,] 9.584596e-02 1.079123e-02 4.232029e-02 5.576051e-01
#> [275,] 2.785985e-02 9.098385e-01 2.112440e-01 2.083367e-01
#> [276,] 3.162166e-02 2.413634e-02 7.686041e-01 4.391543e-03
#> [277,] 1.588926e-01 2.912407e-02 1.917168e-02 4.932617e-01
#> [278,] 1.805632e-02 9.503336e-01 9.952377e-01 1.685768e-02
#> [279,] 3.258032e-01 1.988275e-01 5.570270e-01 2.558491e-01
#> [280,] 3.987206e-01 1.173874e-01 1.586349e-01 3.668415e-01
#> [281,] 1.599052e-01 1.978290e-02 1.153382e-01 3.448234e-01
#> [282,] 2.443178e-05 9.940892e-01 4.978129e-04 1.747556e-02
#> [283,] 1.610094e-02 3.301935e-01 8.896451e-01 2.534227e-02
#> [284,] 5.297566e-01 1.807366e-01 1.062568e-01 3.480824e-01
#> [285,] 7.422993e-02 9.917411e-01 4.929735e-03 4.996106e-01
#> [286,] 9.460792e-02 2.099794e-01 3.364141e-01 2.183551e-01
#> [287,] 7.773471e-01 1.396490e-01 3.111575e-01 4.003077e-01
#> [288,] 5.687328e-03 1.430557e-01 9.935101e-01 1.163073e-02
#> [289,] 3.562498e-02 1.582390e-01 2.954206e-01 8.512366e-01
#> [290,] 2.164325e-01 2.819940e-01 8.113302e-01 5.933186e-03
#> [291,] 2.599485e-01 3.073599e-01 8.725626e-01 7.815084e-02
#> [292,] 4.786031e-02 3.218405e-01 2.306733e-01 2.133850e-01
#> [293,] 4.407082e-01 1.339544e-01 9.954820e-02 6.605424e-02
#> [294,] 6.937388e-01 9.382630e-01 3.475063e-01 4.391164e-01
#> [295,] 3.608294e-02 2.279196e-02 6.481584e-02 8.009488e-02
#> [296,] 3.759409e-01 6.771793e-05 1.224313e-01 6.538978e-03
#> [297,] 1.454753e-01 2.212803e-01 4.413530e-01 1.319301e-01
#> [298,] 4.318884e-01 1.453023e-01 6.497250e-01 2.222865e-01
#> [299,] 9.149777e-01 1.906699e-01 7.876271e-02 1.953318e-01
#> [300,] 2.329272e-01 1.604257e-01 9.331594e-01 1.908563e-02
#> [301,] 4.635561e-01 4.368368e-01 1.023446e-01 1.750690e-01
#> [302,] 2.532577e-02 8.593303e-01 2.005484e-01 3.202274e-02
#> [303,] 9.505196e-03 4.291223e-01 6.390272e-03 6.005856e-01
#> [304,] 9.051342e-02 8.418900e-01 2.243847e-02 8.303714e-02
#> [305,] 5.427900e-01 2.939164e-01 5.312504e-01 2.622170e-01
#> [306,] 9.394853e-01 1.092422e-03 6.250587e-01 1.870801e-02
#> [307,] 6.257512e-01 3.549356e-01 8.670560e-02 9.453414e-02
#> [308,] 9.747777e-01 4.033661e-02 9.398052e-01 2.541398e-01
#> [309,] 6.127867e-01 1.358226e-01 8.825425e-01 7.913635e-02
#> [310,] 7.815524e-01 6.405950e-03 4.165007e-01 1.516676e-01
#> [311,] 9.092764e-01 1.779437e-01 9.783441e-01 6.997373e-02
#> [312,] 1.661176e-01 1.727900e-01 9.466984e-02 7.889214e-01
#> [313,] 9.870216e-01 1.930619e-02 3.311526e-02 4.664307e-01
#> [314,] 1.210280e-01 3.345306e-01 2.666321e-02 3.036575e-01
#> [315,] 9.758732e-01 1.176986e-01 8.639336e-01 8.874085e-01
#> [316,] 1.350756e-01 7.323542e-02 4.097579e-01 1.663239e-01
#> [317,] 9.807465e-01 2.544411e-01 7.787030e-02 8.739376e-01
#> [318,] 4.253529e-02 9.840789e-01 1.263009e-02 2.681441e-01
#> [319,] 2.192698e-01 2.644551e-01 4.522039e-01 1.249149e-01
#> [320,] 8.839571e-01 1.448337e-01 3.382233e-01 7.166038e-02
#> [321,] 2.248985e-01 4.282362e-01 7.636282e-02 1.564799e-01
#> [322,] 2.801691e-01 3.264476e-01 1.771411e-01 9.396606e-01
#> [323,] 3.809275e-01 2.021692e-01 3.803627e-01 3.269293e-02
#> [324,] 3.116181e-01 6.439088e-01 1.848146e-02 8.358883e-01
#> [325,] 1.454106e-01 1.911682e-03 9.174972e-01 3.788830e-01
#> [326,] 8.304356e-01 1.369615e-01 1.183457e-02 2.047323e-01
#> [327,] 3.229442e-02 3.398311e-01 4.248070e-01 7.526020e-01
#> [328,] 1.518108e-03 3.724274e-02 9.977698e-01 3.452140e-03
#> [329,] 3.269934e-02 1.777009e-01 8.287141e-01 1.475104e-01
#> [330,] 2.209584e-01 9.621768e-01 3.132473e-02 3.738215e-01
#> [331,] 9.339075e-02 1.050238e-02 2.375850e-01 3.281991e-01
#> [332,] 4.562627e-01 8.233987e-01 2.682984e-02 1.484343e-01
#> [333,] 9.360276e-02 1.694122e-01 5.881523e-01 6.870955e-02
#> [334,] 2.090761e-01 1.823810e-01 5.612131e-01 1.584282e-01
#> [335,] 2.528643e-01 7.291847e-01 4.968453e-01 6.734563e-02
#> [336,] 3.993974e-02 9.410075e-01 1.798000e-01 1.930265e-01
#> [337,] 2.925196e-01 6.750462e-01 3.253419e-01 3.289223e-01
#> [338,] 3.130699e-02 9.436021e-01 7.470674e-02 5.873670e-03
#> [339,] 9.282657e-01 1.333680e-03 2.334699e-03 9.982068e-02
#> [340,] 2.315669e-02 2.769953e-01 9.895236e-01 1.555567e-04
#> [341,] 9.288233e-02 3.280348e-01 5.215949e-01 2.714966e-01
#> [342,] 1.587896e-01 7.734948e-01 8.147960e-02 9.597140e-02
#> [343,] 4.361858e-01 2.613655e-01 5.119031e-02 1.731097e-01
#> [344,] 4.879346e-02 5.112964e-01 8.152668e-02 6.320693e-01
#> [345,] 5.538104e-01 3.145209e-01 3.829399e-01 2.379997e-01
#> [346,] 2.173045e-02 2.165427e-01 1.553329e-01 8.108791e-01
#> [347,] 2.746540e-01 1.661807e-01 2.378700e-02 6.700761e-01
#> [348,] 9.930514e-01 4.236526e-02 7.185616e-03 5.397096e-01
#> [349,] 8.239969e-02 1.440457e-02 3.569037e-01 8.731776e-01
#> [350,] 4.146859e-04 5.042583e-01 9.225229e-03 7.034632e-01
#> [351,] 2.032695e-01 3.132848e-02 1.645088e-03 9.960251e-01
#> [352,] 9.866265e-01 1.757518e-02 1.846851e-01 2.160831e-01
#> [353,] 7.761482e-01 5.559371e-03 9.337762e-01 7.940022e-02
#> [354,] 2.794856e-01 1.033473e-01 7.351042e-01 8.098210e-01
#> [355,] 7.205819e-01 2.027500e-01 8.659802e-01 4.390092e-01
#> [356,] 4.291152e-01 4.324628e-03 7.128704e-01 2.989633e-02
#> [357,] 5.655941e-02 8.816610e-01 1.251549e-01 1.354389e-02
#> [358,] 1.780573e-03 9.939627e-01 1.170296e-01 2.949506e-02
#> [359,] 3.583535e-01 4.988804e-02 4.600551e-02 1.096940e-01
#> [360,] 1.252008e-01 8.342348e-01 1.192703e-01 1.458119e-01
#> [361,] 9.320588e-01 3.203473e-01 5.667894e-01 1.186606e-01
#> [362,] 8.765604e-01 5.456229e-01 8.538455e-01 4.387630e-01
#> [363,] 4.316130e-01 2.483258e-02 7.271023e-02 9.427172e-01
#> [364,] 1.032704e-02 3.337913e-03 9.997933e-01 1.003202e-04
#> [365,] 6.172159e-03 1.007128e-01 9.114281e-01 4.109194e-01
#> [366,] 6.856282e-01 2.744865e-01 1.843233e-01 3.627841e-01
#> [367,] 1.476520e-02 7.678250e-03 4.590153e-05 4.901974e-01
#> [368,] 4.929739e-01 4.813965e-02 5.677907e-01 1.083074e-01
#> [369,] 7.490940e-02 1.432095e-01 1.407996e-03 6.311892e-01
#> [370,] 2.044692e-01 8.244451e-01 7.699243e-01 9.607534e-03
#> [371,] 2.303221e-02 9.595651e-01 1.737636e-02 5.911857e-02
#> [372,] 3.224383e-01 2.269974e-01 9.833287e-01 5.811882e-02
#> [373,] 7.647231e-04 1.930299e-03 9.833282e-06 9.892936e-01
#> [374,] 4.087356e-01 2.095767e-01 3.453724e-01 1.932101e-01
#> [375,] 3.106367e-01 1.087393e-01 5.189887e-01 1.813856e-01
#> [376,] 1.639437e-02 1.334240e-01 2.681715e-01 2.054218e-01
#> [377,] 6.661307e-02 1.934691e-02 9.781325e-01 5.879738e-01
#> [378,] 3.934134e-01 2.905038e-01 3.710025e-01 1.617192e-01
#> [379,] 4.977497e-01 2.155606e-01 2.736432e-01 4.325650e-01
#> [380,] 3.683248e-01 4.039710e-02 9.370233e-01 3.182732e-01
#> [381,] 1.338303e-01 2.986192e-02 9.799995e-01 8.808677e-03
#> [382,] 9.453388e-01 2.859107e-04 3.666702e-02 8.720095e-01
#> [383,] 1.686165e-02 4.448038e-02 3.458296e-01 2.011714e-02
#> [384,] 5.187580e-01 6.607366e-01 7.030547e-03 9.815634e-01
#> [385,] 2.913876e-02 8.481591e-01 6.338181e-02 7.484045e-01
#> [386,] 2.220355e-02 9.495348e-01 4.080063e-02 4.718564e-01
#> [387,] 3.495774e-01 2.510048e-02 9.789783e-01 1.636402e-02
#> [388,] 4.146041e-03 2.402465e-02 5.839965e-01 4.558290e-01
#> [389,] 1.597316e-02 9.745376e-01 5.520958e-01 2.520979e-03
#> [390,] 3.423940e-02 4.938944e-01 4.297712e-01 3.334513e-02
#> [391,] 7.570964e-03 6.965046e-01 1.592531e-01 5.364511e-02
#> [392,] 7.683642e-02 9.272230e-01 3.622869e-01 1.732113e-01
#> [393,] 2.442515e-01 8.450628e-01 6.126527e-01 5.053123e-01
#> [394,] 8.385814e-01 4.657566e-01 2.608938e-01 4.398809e-02
#> [395,] 1.169648e-01 9.816753e-01 4.426014e-01 3.571548e-01
#> [396,] 6.549175e-01 1.242221e-03 1.609657e-01 1.242029e-02
#> [397,] 1.781168e-02 9.978277e-01 3.480364e-02 5.697682e-02
#> [398,] 4.885353e-01 1.169856e-01 7.128295e-02 8.743186e-01
#> [399,] 1.903868e-01 3.931184e-03 1.934298e-03 7.766157e-01
#> [400,] 3.253608e-03 5.643617e-02 4.057887e-03 9.973841e-01
#> [401,] 5.809006e-01 9.776870e-01 2.074411e-01 9.915057e-01
#> [402,] 2.432504e-01 1.244951e-01 4.589367e-01 8.799927e-01
#> [403,] 8.107641e-01 6.954520e-02 2.120875e-01 6.024754e-02
#> [404,] 8.457358e-01 4.292980e-02 6.194497e-01 2.757129e-02
#> [405,] 1.559952e-03 9.992100e-01 8.503856e-01 1.302482e-03
#> [406,] 2.398170e-01 2.574975e-01 7.764940e-01 4.534261e-01
#> [407,] 5.133278e-01 1.109977e-02 3.867633e-02 9.935700e-01
#> [408,] 1.034006e-01 6.863778e-01 9.479241e-02 1.537168e-01
#> [409,] 1.857159e-01 9.624338e-02 1.301625e-01 5.257187e-01
#> [410,] 2.458143e-03 1.844488e-01 9.634571e-01 2.724071e-02
#> [411,] 8.101042e-01 2.923149e-02 2.011778e-02 7.831567e-01
#> [412,] 9.593556e-01 9.853946e-01 2.128670e-02 2.726820e-01
#> [413,] 2.379272e-01 6.322425e-01 6.928677e-01 3.826769e-01
#> [414,] 3.154628e-02 1.433572e-01 4.081572e-01 1.309240e-02
#> [415,] 5.040513e-02 8.802207e-01 6.779862e-02 2.166411e-02
#> [416,] 8.568439e-02 3.690189e-02 9.464802e-01 1.082177e-01
#> [417,] 4.404967e-01 3.594872e-01 6.024905e-02 6.987077e-01
#> [418,] 9.175493e-01 5.923043e-03 2.007958e-01 5.986722e-01
#> [419,] 2.208141e-02 4.146824e-01 9.693696e-01 5.414226e-02
#> [420,] 6.083357e-02 1.885836e-01 9.799587e-01 8.133425e-03
#> [421,] 1.073133e-01 6.104815e-01 7.724123e-02 3.007659e-01
#> [422,] 2.584951e-02 1.818086e-02 9.683790e-01 4.387788e-03
#> [423,] 1.263316e-01 7.504683e-01 9.756217e-01 3.644788e-01
#> [424,] 9.891042e-01 2.679250e-04 1.073246e-02 4.176938e-05
#> [425,] 8.469593e-01 7.539167e-01 1.407161e-01 6.482294e-02
#> [426,] 5.277591e-01 5.178666e-02 6.511481e-02 2.693605e-01
#> [427,] 9.407212e-01 6.569006e-03 2.185216e-01 9.786654e-03
#> [428,] 1.465513e-01 4.285455e-01 4.815019e-01 4.213288e-02
#> [429,] 2.491049e-01 5.534787e-01 4.488395e-01 9.562250e-01
#> [430,] 3.924627e-01 1.297375e-01 8.179546e-01 9.201944e-02
#> [431,] 7.685030e-01 2.922005e-03 6.027585e-02 1.282085e-03
#> [432,] 7.901665e-02 3.557549e-01 1.689243e-01 1.602263e-01
#> [433,] 8.906521e-02 1.545134e-01 3.333581e-01 2.209279e-01
#> [434,] 2.758939e-01 2.263574e-01 4.652603e-01 4.812269e-02
#> [435,] 2.716038e-02 3.580544e-03 9.718019e-01 9.380615e-02
#> [436,] 3.792564e-01 8.818463e-02 1.552795e-01 6.787641e-01
#> [437,] 4.324253e-01 8.303097e-02 6.203604e-01 2.241529e-01
#> [438,] 7.456188e-02 2.172775e-02 1.263105e-01 5.000475e-01
#> [439,] 2.216436e-02 4.231027e-01 2.592555e-01 2.637973e-01
#> [440,] 2.316900e-02 9.737603e-01 1.404741e-01 6.199878e-01
#> [441,] 1.795917e-02 6.158462e-02 8.566817e-01 7.234913e-02
#> [442,] 1.382535e-03 3.518627e-01 5.714405e-01 1.762687e-03
#> [443,] 1.327153e-01 2.668062e-02 9.970934e-01 6.741217e-04
#> [444,] 7.092377e-02 7.404882e-01 4.408806e-01 1.362275e-01
#> [445,] 1.112339e-02 9.775115e-01 1.712059e-01 9.167719e-01
#> [446,] 8.445795e-01 6.791698e-01 2.009874e-02 8.017069e-01
#> [447,] 6.958203e-01 1.280623e-04 3.047096e-03 4.427388e-02
#> [448,] 1.548138e-01 3.453124e-01 8.712427e-01 2.818296e-01
#> [449,] 1.676835e-01 9.950847e-01 1.659787e-02 7.791344e-01
#> [450,] 1.026987e-01 1.060509e-01 3.625900e-01 9.939443e-01
#> [451,] 4.591747e-01 3.279373e-01 1.920137e-01 6.553159e-01
#> [452,] 1.125667e-02 1.320240e-01 5.088361e-01 1.062128e-02
#> [453,] 9.961257e-01 7.497593e-02 4.292593e-01 2.255328e-01
#> [454,] 9.813526e-01 7.732340e-02 5.212850e-02 6.524487e-02
#> [455,] 9.546930e-01 3.076089e-02 1.582013e-01 1.532292e-02
#> [456,] 4.808501e-01 1.034375e-01 2.228830e-01 3.688453e-01
#> [457,] 2.068733e-01 2.736928e-01 4.105142e-01 2.497445e-02
#> [458,] 9.627487e-01 1.243955e-02 1.917393e-01 1.359390e-02
#> [459,] 3.999172e-01 2.213288e-01 5.810469e-01 1.057926e-01
#> [460,] 9.811733e-02 3.914780e-01 9.812016e-01 4.459865e-03
#> [461,] 6.606237e-01 8.517118e-01 8.351589e-03 4.907868e-02
#> [462,] 7.948205e-05 2.758613e-01 2.710816e-02 2.121449e-03
#> [463,] 3.347872e-01 8.189757e-02 5.608196e-01 4.284817e-01
#> [464,] 5.563710e-01 9.475482e-02 6.416030e-01 5.601751e-01
#> [465,] 9.648813e-01 3.457625e-02 7.743672e-02 9.521829e-02
#> [466,] 4.137836e-01 7.245581e-05 2.180935e-02 1.143351e-03
#> [467,] 1.973769e-01 5.666278e-02 8.058757e-01 1.774290e-01
#> [468,] 1.143196e-01 8.951316e-01 4.577544e-02 4.904266e-02
#> [469,] 3.826511e-01 5.333501e-01 3.742088e-01 1.615507e-01
#> [470,] 9.515017e-01 9.742701e-01 5.624188e-01 6.448420e-01
#> [471,] 6.570135e-01 7.103939e-01 6.121082e-01 8.514040e-03
#> [472,] 3.416046e-01 4.318347e-01 2.381594e-02 1.078069e-02
#> [473,] 1.153417e-02 1.282129e-01 3.802015e-01 6.084209e-01
#> [474,] 5.979535e-01 2.037364e-01 3.796632e-02 6.991171e-01
#> [475,] 2.947253e-01 9.772239e-04 5.800558e-01 3.588462e-01
#> [476,] 9.994256e-01 5.089454e-04 7.460476e-04 2.387097e-02
#> [477,] 3.768765e-01 7.947169e-01 1.456006e-01 3.223668e-02
#> [478,] 1.460744e-02 6.129705e-01 3.435659e-02 3.695443e-01
#> [479,] 2.074102e-01 4.226421e-01 4.393706e-01 5.336313e-02
#> [480,] 3.392307e-01 9.662457e-01 6.409889e-02 3.418814e-01
#> [481,] 4.233649e-03 3.449481e-01 8.806166e-03 6.225114e-01
#> [482,] 6.216002e-01 9.112186e-02 4.603337e-02 1.843262e-01
#> [483,] 3.835825e-01 1.320383e-02 6.394834e-01 3.588270e-01
#> [484,] 2.717029e-01 3.326384e-03 2.333754e-03 7.165271e-01
#> [485,] 2.732443e-01 2.941614e-01 3.123349e-01 4.864241e-01
#> [486,] 2.047618e-01 1.663323e-02 3.502312e-01 1.307597e-01
#> [487,] 5.184493e-01 7.476824e-02 9.383563e-01 5.740971e-01
#> [488,] 5.563590e-02 9.686907e-01 9.393800e-01 1.902869e-01
#> [489,] 6.778181e-03 9.575875e-01 9.449692e-01 5.081901e-01
#> [490,] 6.300604e-02 3.557221e-01 2.493623e-02 9.867850e-01
#> [491,] 7.556657e-01 1.292109e-01 2.617469e-01 5.668969e-01
#> [492,] 1.568600e-01 3.163450e-02 1.457157e-02 5.794092e-02
#> [493,] 4.439653e-01 7.205638e-01 1.251768e-01 2.059731e-01
#> [494,] 9.950908e-01 6.278961e-02 1.466003e-01 1.071455e-03
#> [495,] 9.047119e-02 4.584428e-01 2.267621e-01 1.161120e-01
#> [496,] 8.076507e-01 3.402474e-01 7.910072e-01 2.877535e-01
#> [497,] 8.335849e-01 5.894141e-02 4.995519e-01 3.630861e-01
#> [498,] 7.727277e-01 5.953971e-02 2.090493e-01 2.712673e-01
#> [499,] 4.040443e-01 7.901651e-01 4.317461e-01 3.892217e-01
#> [500,] 1.559961e-01 5.523894e-01 3.029844e-02 5.452396e-01
#> 
#> $fits[[1]]$perf
#>         1         2         3         4 
#> 0.3317149 0.3088877 0.3667880 0.2984700 
#> 
#> $fits[[1]]$matrix
#>           1         2         3         4
#> 1 165.85747 157.97111 100.35328  75.81813
#> 2 167.57558 154.44387  93.73873  84.24182
#> 3  82.33653  74.15231 183.39401 160.11715
#> 4  86.24874  83.97552 180.54073 149.23501
#> 
#> $fits[[1]]$sse
#> [1] 0.1232035
#> 
#> $fits[[1]]$data
#> xslData object with label "example condition" and condition "ambiguous"
#>   training trials: 2
#>       test trials: 0
#>             words: 4
#>           objects: 4
#>        accuracies: 4
#> 
#> 
#> 
#> $sse
#> [1] 0.1232035
#> 
#> $unweighted_sse
#> [1] 0.1232035
#>